requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pandas>=1.3.0
google-analytics-data>=0.16.0
google-auth>=2.3.0
openai>=1.0.0
httpx>=0.23.0
gspread>=5.7.0
litellm>=1.0.0
python-dotenv>=0.19.0
python-dateutil>=2.8.2
